topic4/requirements_travel_planner.txt
# Smart Travel Planner - Python Dependencies
# Install with: pip install -r requirements_travel_planner.txt

# Core LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20

# HTTP requests for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0



